refactor(#3631): ONE require_fully_consumed, per the ruling recorded on main
`raw_value/reporting.rs` recorded the obligation in its own source: PR #3820 was
adding `typed_value.rs::require_fully_consumed`, "there must be ONE
implementation of this rule, not two: when #3820 lands, this function becomes a
one-line delegation to it (or is deleted in favour of it) and the two error
message classes are unified. Fold in at the same time `cell_path_key.rs`'s
now-redundant `duration` consumption check."

Taken in the DELETE direction rather than the delegate direction: two names for
one rule is what a future author picks between by accident. So

* `require_fully_consumed_raw` is GONE. The one implementation lives in the new
  `typed_value.rs` as `require_fully_consumed`, with the SAME signature and the
  SAME message, so the type-STRING side and the (incoming) `&CqlType` side share
  one error class. All seven call sites — `raw_value.rs`'s bounded wrapper,
  `cell_value_complex.rs` x2, `udt.rs` x3, `udt/inline.rs` — name it unchanged.
* `cell_path_key.rs`'s `duration` consumption walk is folded out. It measured the
  three-VInt framing itself and returned `Some(pos)`, which was live only while
  `parse_value_from_raw_bytes` discarded its own count. #3811 made that function
  assert exhaustion and made its `"duration"` arm report where the third VInt
  ended, so the callee refuses a trailing byte one layer down and the walk here
  could never fire. Its refusal is the same "decoded only N of M byte(s)" class,
  which is why `cell_path_key_tests`'s pinned assertions are unaffected
  (44 passed, including `duration_cell_path_key_rejects_trailing_bytes`).

Also splits `row_decoder/mod.rs`'s `UnfilteredSerializer` flag constants into
`row_flags.rs` (epic #1116): that file is over the 800-line source target, and it
has to gain the `mod typed_value;` declaration. Net 877 -> 861 lines.

Refs #3631, #3811.
