+++ cqlite-core/src/storage/sstable/reader/parsing/row_decoder/raw_value.rs
@@ -104,7 +104,7 @@ impl V5CompressedLegacyParser {
     /// [`Self::parse_value_from_raw_bytes_reporting`], which threads a REAL
     /// consumption count out of every arm; the wrapper then requires the decode to
     /// have consumed every byte of `data`
-    /// ([`Self::require_fully_consumed_raw`]). The rule is
+    /// ([`Self::require_fully_consumed`]). The rule is
     /// `cassandra-5.0.8` `TupleType.split`: a genuinely SHORT encoding leaves
     /// `position == length` and is legal, while trailing bytes (rule 4) and a
     /// partial 1-3 byte component-length prefix (rule 2) both leave it short and
@@ -127,7 +127,7 @@ impl V5CompressedLegacyParser {
     ) -> Result<Value> {
         let (value, consumed) =
             self.parse_value_from_raw_bytes_reporting(data, type_str, column_name, depth)?;
-        Self::require_fully_consumed_raw(consumed, data.len(), column_name, type_str)?;
+        Self::require_fully_consumed(consumed, data.len(), column_name, type_str)?;
         Ok(value)
     }
 }
